requirements.txt
python-telegram-bot[job-queue]==20.7
python-telegram-bot[rate-limiter]==20.7 